+++ src/main/java/com/backend/ClinicaOdontologica/service/IPacienteService.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface IPacienteService {
     PacienteSalidaDto registrarPaciente(PacienteEntradaDto pacienteEntradaDto);
     List<PacienteSalidaDto> listarPacientes();
-    PacienteSalidaDto buscarPorId(int id);
+    PacienteSalidaDto buscarPorId(Long id);
 }
